+++ libcaf_core/caf/actor_pool.cpp
@@ -7,6 +7,7 @@
 #include "caf/anon_mail.hpp"
 #include "caf/default_attachable.hpp"
 #include "caf/detail/assert.hpp"
+#include "caf/detail/current_actor.hpp"
 #include "caf/detail/sync_request_bouncer.hpp"
 #include "caf/mailbox_element.hpp"
 
@@ -122,7 +123,7 @@ const char* actor_pool::name() const {
 }
 
 void actor_pool::on_cleanup([[maybe_unused]] const error& reason) {
-  CAF_PUSH_AID_FROM_PTR(this);
+  detail::current_actor_guard ctx_guard{this};
   CAF_LOG_TERMINATE_EVENT(this, reason);
 }